dentalAI/requirements.txt
django
numpy
Pillow
django-nextjs
tensorflow
psycopg2-binary
django-cors-headers
langchain-openai
langchain
chroma
langchain-community
langchain-openai
python-dotenv
llama-index
torch
transformers
pytesseract